requirements.txt
requests>=2.25.1
click>=8.0.0
python-dotenv>=0.19.0
bitfinex-api-py>=1.1.8
rich>=13.0.0

# Clean Architecture dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
dependency-injector>=4.41.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
fakeredis>=2.20.0
pytest-mock>=3.10.0

# Additional utilities
tenacity>=8.2.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4